make parseRouter serialisable v2 (updated 09:30)
